before_action のメソッドを共通化
